# frontend-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateEntryRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "kind": {
            "enum": [
              "file",
              "directory"
            ],
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "kind"
        ],
        "title": "CreateEntryRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "The common wire format for API errors.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FileContent": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "modifiedAt": {
            "format": "date-time",
            "title": "Modifiedat",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "revision",
          "size",
          "modifiedAt",
          "content"
        ],
        "title": "FileContent",
        "type": "object"
      },
...
      "FileEntryResponse": {
        "properties": {
          "kind": {
            "enum": [
              "file",
              "directory"
            ],
            "title": "Kind",
            "type": "string"
          },
          "modifiedAt": {
            "format": "date-time",
            "title": "Modifiedat",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "required": [
          "path",
          "name",
          "kind",
          "size",
          "modifiedAt",
          "revision"
        ],
        "title": "FileEntryResponse",
        "type": "object"
      },
...
      "RenameEntryRequest": {
        "additionalProperties": false,
        "properties": {
          "baseRevision": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Baserevision",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "destination",
          "baseRevision"
        ],
        "title": "RenameEntryRequest",
        "type": "object"
      },
...
      "SaveFileRequest": {
        "additionalProperties": false,
        "properties": {
          "baseRevision": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Baserevision",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "baseRevision",
          "content"
        ],
        "title": "SaveFileRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ControlLeaseToken": {
        "in": "header",
        "name": "X-Control-Lease-Token",
        "type": "apiKey"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "manyselves_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/files/content": {
      "get": {
        "operationId": "get_api_v1_projects_project_id_files_content",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Read File"
      },
      "put": {
        "operationId": "put_api_v1_projects_project_id_files_content",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Save File"
      }
    },
    "/api/v1/projects/{project_id}/files/download": {
      "get": {
        "operationId": "get_api_v1_projects_project_id_files_download",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Download File"
      }
    },
    "/api/v1/projects/{project_id}/files/entries": {
      "delete": {
        "operationId": "delete_api_v1_projects_project_id_files_entries",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Delete Entry"
      },
      "post": {
        "operationId": "post_api_v1_projects_project_id_files_entries",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Create Entry"
      }
    },
    "/api/v1/projects/{project_id}/files/preview": {
      "get": {
        "operationId": "get_api_v1_projects_project_id_files_preview",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Preview File"
      }
    },
    "/api/v1/projects/{project_id}/files/rename": {
      "post": {
        "operationId": "post_api_v1_projects_project_id_files_rename",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Rename Entry"
      }
    },
    "/api/v1/projects/{project_id}/files/tree": {
      "get": {
        "operationId": "get_api_v1_projects_project_id_files_tree",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "",
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileTreeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "File Tree"
      }
    },
    "/api/v1/projects/{project_id}/files/upload": {
      "post": {
        "operationId": "post_api_v1_projects_project_id_files_upload",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Upload File"
      }
    },
...
  }
}